--- app/models/base_host.rb
@@ -1,5 +1,0 @@
-class BaseHost < ActiveRecord::Base
-
-  self.table_name = "hosts"
-
-end
