+++ app/models/host/base.rb
@@ -1,0 +1,9 @@
+require 'facts_importer'
+
+module Host
+  class Base < ActiveRecord::Base
+
+    self.table_name = "hosts"
+
+  end
+end
