+++ app/models/host_observer.rb
@@ -1,7 +1,9 @@
 class HostObserver < ActiveRecord::Observer
 
+  observe Host::Base
+
   # Sets and expire provisioning tokens
-  # this has to happen post validation and before the orchesration queue is starting to 
+  # this has to happen post validation and before the orchesration queue is starting to
   # process, as the token value is required within the tftp config file manipulations
   def after_validation(host)
     return unless SETTINGS[:unattended]
